produtos e associacao

+++ spring-boot-ionic_nelioAlves/projeto_do_curso/curso_rest/src/main/java/com/projetoNelioAlves/curso_rest/resources/CategoriaResource.java
@@ -22,7 +22,7 @@ public class CategoriaResource {
     private CategoriaService categoriaService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> find(@PathVariable long id) {
+    public ResponseEntity<?> find(@PathVariable Integer id) {
 
         return ResponseEntity.ok().body(categoriaService.buscar(id));
     }
