+++ spring-boot-ionic_nelioAlves/projeto_do_curso/curso_rest/src/main/java/com/projetoNelioAlves/curso_rest/service/CategoriaService.java
@@ -15,7 +15,7 @@ public class CategoriaService {
 
     }
 
-    public Categoria buscar(long id){
+    public Categoria buscar(Integer id){
       return  categoriaRepository.findById(id).get();
     }
